requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==1.26.4
tqdm==4.66.5
pytest==8.3.3
pillow==10.4.0
